A commit motivated by SunOS veteran Winston. He gave SunOS 4.1.2 with gcc 2.7.0 a spin.
A commit that looks bigger than it is. Most files only got header #includes removed that are handled in mpg123.h .
Then there are the corresponding changes in mpg123.h; together with the addition of safe_realloc for ancient systems.
The ugly need for snprintf in make_remote_header() has been avoided by changing this one into print_remote_header().
The Question of size_t, ssize_t and off_t is handled in configure now; respectively in config.h.legacy .



git-svn-id: svn://scm.orgis.org/mpg123/trunk@861 35dc7657-300d-0410-a2e5-dc2837fedb53