Restructure the synth function choosing; removing a good deal of ugly code.
After realizing that the synth function calls are via the pointer in fr->synth (usually), the opt_synth_ macros have been eliminated.
The frame struct now hosts a struct with arrays of all synth functions that are available for the chosen decoder -- instead of lots if individual pointer variables. That way, the code in optimize.c could be simplified.

Also, the changes for stereo synths are hopefully small, and even the i486 code should then be able to be integrated in a x86 build; though one can debate on how much sense this makes.



git-svn-id: svn://scm.orgis.org/mpg123/trunk@1945 35dc7657-300d-0410-a2e5-dc2837fedb53